Rename 'command' to 'on'
